Separate market navigation to individual component
